KW-337/chore: mobile, admin 컨트롤러 분리
