+++ src/main/java/com/doubleo/hospitalservice/domain/hospital/controller/HospitalAdminController.java
@@ -3,7 +3,7 @@ package com.doubleo.hospitalservice.domain.hospital.controller;
 import com.doubleo.hospitalservice.domain.hospital.dto.response.HospitalInfoResponse;
 import com.doubleo.hospitalservice.domain.hospital.service.HospitalService;
 import io.swagger.v3.oas.annotations.Operation;
-import java.util.List;
+import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -13,16 +13,11 @@ import org.springframework.web.bind.annotation.RestController;
 @RestController
 @RequestMapping("/hospitals")
 @RequiredArgsConstructor
-public class HospitalController {
+public class HospitalAdminController {
 
     private final HospitalService hospitalService;
 
-    @GetMapping
-    @Operation(summary = "All Hospitals get API", description = "모든 병원을 조회하기 위한 API")
-    public List<HospitalInfoResponse> hospitalListGetAll() {
-        return hospitalService.getAllHospitals();
-    }
-
+    @Tag(name = "A-4. Hospital API", description = "병원 관련 API")
     @GetMapping(path = "/{hospitalId}")
     @Operation(summary = "Hospital Detail get API", description = "병원 상세 정보를 조회하기 위한 API")
     public HospitalInfoResponse hospitalDetailGet(@PathVariable Long hospitalId) {
